+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - e2e/cypress tests([#229](https://github.com/boarlabsxyz/Kavoon/pull/229))
+- issues with updating the language selector and checkbox state on the status/reviews page ([#228](https://github.com/boarlabsxyz/Kavoon/pull/228))
 
 ### Changed 
 - the YAML file for deployment([#226](https://github.com/boarlabsxyz/Kavoon/pull/226))
@@ -20,9 +21,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - @next/bundle-analyzer from the Kavoon project ([#223](https://github.com/boarlabsxyz/Kavoon/pull/223))
 - unused prints (dinosaurs, scarlet_ibis, star_flower, foxes, parrots, leaf_on_black, jungle, forest) and their corresponding pictures ([#217](https://github.com/boarlabsxyz/Kavoon/pull/217))
-
-### Fixed
-- issues with updating the language selector and checkbox state on the status/reviews page ([#228](https://github.com/boarlabsxyz/Kavoon/pull/228))
 
 ## [1.3.0] - 2025-02-10